# Conflicts:
#	frontend/src/app/board/board-detail.component.css
#	frontend/src/app/board/board-detail.component.html

+++ frontend/src/app/board/board-detail.component.ts
@@ -1,12 +1,16 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, HostListener } from '@angular/core';
 import { BoardDragDropService } from '../services/board-dragdrop.service'
 import { CdkDragDrop, moveItemInArray } from '@angular/cdk/drag-drop';
 import { DragulaService } from "ng2-dragula";
 import { from, Subscription } from "rxjs";
-import {Router, ActivatedRoute, Params} from '@angular/router';
+import { Router, ActivatedRoute, Params } from '@angular/router';
 import * as _ from 'underscore';
-import {MatDialog, MatDialogRef, MAT_DIALOG_DATA} from '@angular/material/dialog';
+import { MatDialog, MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
 import { ColumnService } from '../services/column.service';
+import { ConfirmDialogComponent } from '../confirm-dialog/confirm-dialog.component';
+
+import { DataSharingService } from '../services/data-sharing-service.service';
+
 
 @Component({
   selector: 'board-detail',
@@ -14,85 +18,68 @@ import { ColumnService } from '../services/column.service';
   styleUrls: ['board-detail.component.css']
 })
 export class BoardDetailComponent implements OnInit {
-  //TODO: add spiner state
-
   content: any
-  isEdited = false;
-  inputAdded = false;
+  editTimeOut: any
+  isEdited = false
+  inputAdded = false
+  boardName: String
 
-  public columns: Array<any> = [
-    {
-    name: 'Todo',
-    cards: [{ title: 'Item A' }, { title: 'Item B' }, { title: 'Item C' }, { title: 'Item D' }]
-    },
-    {
-    name: 'Doing',
-    cards: [{ title: 'Item 1' }, { title: 'Item 2' }, { title: 'Item 3' }, { title: 'Item 4' }]
-    },
-  ];
+  public columns: Array<any> = [];
 
   boardId: any
-
 
   subs = new Subscription();
   constructor(
     private boardDDService: BoardDragDropService,
     private dragulaService: DragulaService,
     private activatedRoute: ActivatedRoute,
+    private dataSharingService: DataSharingService,
     public dialog: MatDialog,
-    private columnService : ColumnService
+    private columnService: ColumnService
   ) {
     this.boardId = this.activatedRoute.snapshot.paramMap.get('id');
-    try{
+    try {
       this.dragulaService.createGroup("COLUMNS", {
         direction: 'horizontal',
         moves: (el, source, handle) => {
           return handle.className === "group-handle"
         }
       });
-    } catch(e){}
- 
+    } catch (e) { }
+
     this.subs.add(this.dragulaService.drop()
       .subscribe(({ name, el, target, source, sibling }) => {
-        console.log("drop catch all ", el ,target, sibling)
-        if(el.getAttribute("id") == null) {
+        console.log('ddrp')
+        if (el.getAttribute("id") == null) {
           return false
         }
-      
+
         let targetColumnId = target.getAttribute('id')
         let childNodes = target.children
-        
-       
-        let oldCardOrder = el.getAttribute("order")
         let movedCardId = el.getAttribute("id")
-        let sourceColumnId = el.getAttribute("data-colid")  
-    
-        if(targetColumnId === sourceColumnId) {
+        let sourceColumnId = el.getAttribute("data-colid")
+
+        if (targetColumnId === sourceColumnId) {
           // change order of cards in a column
           for (let index = 0; index < childNodes.length; index++) {
             const element = childNodes[index];
-            if(element.getAttribute('id')) {
+            if (element.getAttribute('id')) {
               let body = {
-                "column_id" : targetColumnId,
-                "new_order" : index + 1,
-                "card_id" : element.getAttribute('id')
+                "column_id": targetColumnId,
+                "new_order": index + 1,
+                "card_id": element.getAttribute('id')
               }
-  
-              this.boardDDService.changeOrderCard(this.boardId, body)
-              .subscribe(re => console.log(re))
-            } else {
-              console.log('leuleu')
-            }
-          }
 
-          // targetColumnId
-          // boardid
-          // 
+              this.boardDDService.changeOrderCard(this.boardId, body)
+                .subscribe(re => console.log(re))
+            }
+
+          }
         } else {
-         
+
           // getinfo old card
           this.boardDDService.getCardById(this.boardId, sourceColumnId, movedCardId).subscribe(cr => {
-            if(cr['columns']) {
+            if (cr['columns']) {
               let foundedcol = _.filter(cr['columns'], col => {
                 return col._id == sourceColumnId
               })
@@ -108,11 +95,10 @@ export class BoardDetailComponent implements OnInit {
                * delete old card at old column
               */
               this.boardDDService.addCardToColumn(this.boardId, targetColumnId, card[0]).subscribe(ars => {
-                if(ars) {
+                if (ars) {
                   this._updateOrderCard(targetColumnId, childNodes)
                   this.boardDDService.deleteCardOutOfColumn(this.boardId, sourceColumnId, movedCardId).subscribe(dres => console.log(dres))
                 }
-
                 // delete old card
               })
             }
@@ -121,39 +107,25 @@ export class BoardDetailComponent implements OnInit {
       })
     );
 
-     this.subs.add(this.dragulaService.dropModel("COLUMNS")
+    this.subs.add(this.dragulaService.dropModel("COLUMNS")
       .subscribe(({ el, target, source, sourceModel, targetModel, item }) => {
-       for (let index = 0; index < targetModel.length; index++) {
-         const element = targetModel[index];
-         let body = {
-           "column_id" : element._id,
-           "new_order" : index + 1
-         }
-         this.boardDDService.changeOrderColumn(this.boardId, body)
-         .subscribe(re => console.log(re))
-       }
+        for (let index = 0; index < targetModel.length; index++) {
+          const element = targetModel[index];
+          let body = {
+            "column_id": element._id,
+            "new_order": index + 1
+          }
+          this.boardDDService.changeOrderColumn(this.boardId, body)
+            .subscribe(re => console.log(re))
+        }
       })
-    );
+    );  
   }
-
-  _updateOrderCard(targetColumnId, childNodes){
-    for (let index = 0; index < childNodes.length; index++) {
-      const element = childNodes[index];
-      let body = {
-        "column_id" : targetColumnId,
-        "new_order" : index + 1,
-        "card_id" : element.getAttribute('id')
-      }
-
-      this.boardDDService.changeOrderCard(this.boardId, body)
-      .subscribe(re => console.log(re))
-    }
-  }
-
+  
   ngOnInit(): void {
     this.boardDDService.getById(this.boardId).subscribe(
       res => {
-        if (!res ||res['error'] || res["_id"] == null || typeof res["_id"] === "undefined") {
+        if (!res || res['error'] || res["_id"] == null || typeof res["_id"] === "undefined") {
           this.content = "Error " + res['error']
         } else {
           this.content = res
@@ -161,9 +133,24 @@ export class BoardDetailComponent implements OnInit {
             element.cards = _.sortBy(element.cards, "order")
           });
           this.columns = _.sortBy(res['columns'], "order")
+          this.boardName = res['name']
         }
       }
     )
+  }
+
+  _updateOrderCard(targetColumnId, childNodes) {
+    for (let index = 0; index < childNodes.length; index++) {
+      const element = childNodes[index];
+      let body = {
+        "column_id": targetColumnId,
+        "new_order": index + 1,
+        "card_id": element.getAttribute('id')
+      }
+
+      this.boardDDService.changeOrderCard(this.boardId, body)
+        .subscribe(re => console.log(re))
+    }
   }
 
   ngOnDestroy() {
@@ -176,45 +163,75 @@ export class BoardDetailComponent implements OnInit {
 
   // Update cloumn name
   editColumnName(event, col) {
-    console.log(event.target.value)
     let newColumnName = event.target.value
-    let coulmnID = col._id
-    this.columnService.updateColumnName(coulmnID , this.boardId, newColumnName).
-      subscribe(re => console.log(re))
-    
+    if(newColumnName === col.name) {
       col.isEdited = false;
-      col.name = newColumnName
+      return false
+    }
+    let coulmnID = col._id
+    this.columnService.updateColumnName(coulmnID, this.boardId, newColumnName).
+      subscribe(re => console.log(re))
 
-    //TODO: call api to update name column here
-    // then set in the callback like bellow
+    col.isEdited = false;
+    col.name = newColumnName
 
-  
   }
- 
+
+  // Delete a card
+  deleteCardItem(event, card, colId) {
+    const confirmDialog = this.dialog.open(ConfirmDialogComponent, {
+      data: {
+        title: 'Confirm Remove Card',
+        message: 'Are you sure you want to remove this card?'
+      }
+    });
+
+    confirmDialog.afterClosed().subscribe(result => {
+      if (result === true) {
+        let cardIdForDelete = card._id
+        console.log('good to remove ', cardIdForDelete, colId)
+        //TODO: Sami call service to delete card here, and just print out the response console.log.
+      }
+    });
+  }
+
   // Delete a column
   deleteColumn(event, col) {
-    console.log(event.target)
-    console.log(col)
-    let columnIdForDelete = col._id
-    this.columnService.deleteColumn(columnIdForDelete, this.boardId).
-      subscribe(re => console.log(re))
+    const confirmDialog = this.dialog.open(ConfirmDialogComponent, {
+      data: {
+        title: 'Confirm Remove Column',
+        message: 'Are you sure you want to remove this column: ' + col.name
+      }
+    });
 
-    col.isDeleted = true;
+    confirmDialog.afterClosed().subscribe(result => {
+      if (result === true) {
+        let columnIdForDelete = col._id
+        this.columnService.deleteColumn(columnIdForDelete, this.boardId).
+          subscribe(re => console.log(re))
 
-    //TODO: call api to delete column here
-    // then show console.log response.
-    // I will handle UI and make it disappear.
-
+        col.isDeleted = true;
+        this.columns = _.without(this.columns, _.findWhere(this.columns, {
+          _id: col._id
+        }));
+      }
+    });
   }
-  
+
   // Add a new column
   addNewColumn(event) {
-    console.log(event.target.value)
     let newColumnName = event.target.value
     this.columnService.addNewColumn(this.boardId, newColumnName).
-      subscribe(re => console.log(re))
+      subscribe(re => {
+        this.columns.push(re)
+        event.target.value = ""
+      })
+  }
 
-    //TODO: Tung Note. call api to create new column here
-    // And dont handle about callback UI, I will take it
+  addNewCard(event) {
+    var match = _.find(this.columns, function(item) { return item._id === event.col_id })
+    if (match) {
+        match.cards.push(event)
+    }
   }
 }
